Move lots of things around, start adding tests
