+++ .idea/infra.iml
@@ -9,6 +9,7 @@
       <excludeFolder url="file://$MODULE_DIR$/.env" />
       <excludeFolder url="file://$MODULE_DIR$/.mypy_cache" />
       <excludeFolder url="file://$MODULE_DIR$/.pytest_cache" />
+      <excludeFolder url="file://$MODULE_DIR$/.poetry" />
     </content>
     <orderEntry type="jdk" jdkName="Poetry (infra)" jdkType="Python SDK" />
     <orderEntry type="sourceFolder" forTests="false" />
